feat: Refactor to retain fromXXX methods.
