+++ src/format/from-csv.js
@@ -1,12 +1,8 @@
-import { ColumnTable } from '../table/ColumnTable.js';
-import { error } from '../util/error.js';
-import { isString } from '../util/is-string.js';
-import { DelimitedTextStream } from './stream/delimited-text-stream.js';
-import { lineFilter } from './stream/line-filter-stream.js';
-import { parseTextRows } from './stream/parse-text-rows.js';
-import { pipeline } from './stream/pipeline.js';
+import { ColumnTable } from '../table/ColumnTable.js'; // eslint-disable-line no-unused-vars
+import { delimitedTextTransformer } from './stream/delimited-text-stream.js';
+import { lineFilterTransformer } from './stream/line-filter-stream.js';
+import { parseTextRowsStream, parseTextRowsSync } from './stream/parse-text-rows.js';
 import { textStream } from './stream/text-stream.js';
-import { toTextStream } from './stream/to-text-stream.js';
 
 /**
  * Options for CSV parsing.
@@ -31,16 +27,19 @@ import { toTextStream } from './stream/to-text-stream.js';
  */
 
 /**
- * Load a CSV file from a URL and return a Promise for an Arquero table.
- * @param {string} path The URL or file path to load.
- * @param {import('./types.js').LoadOptions & CSVParseOptions} [options]
- *  CSV parse options.
- * @return {Promise<ColumnTable>} A Promise to an Arquero table.
- * @example aq.loadCSV('data/table.csv')
- * @example aq.loadTSV('data/table.tsv', { delimiter: '\t' })
+ * Parse a comma-separated values (CSV) string into a table. Other
+ * delimiters, such as tabs or pipes ('|'), can be specified using
+ * the options argument. By default, automatic type inference is performed
+ * for input values; string values that match the ISO standard
+ * date format are parsed into JavaScript Date objects. To disable this
+ * behavior, set the autoType option to false. To perform custom parsing
+ * of input column values, use the parse option.
+ * @param {string} input The input text.
+ * @param {CSVParseOptions} [options] The formatting options.
+ * @return {ColumnTable} An Arquero table.
  */
-export async function loadCSV(path, options) {
- return parseCSV(await textStream(path, options), options);
+export function fromCSV(input, options) {
+  return parseTextRowsSync(input, transforms(options), options);
 }
 
 /**
@@ -51,29 +50,34 @@ export async function loadCSV(path, options) {
  * date format are parsed into JavaScript Date objects. To disable this
  * behavior, set the autoType option to false. To perform custom parsing
  * of input column values, use the parse option.
- * @param {ReadableStream<string> | string} input The input text or stream.
+ * @param {ReadableStream<string>} stream The input stream.
  * @param {CSVParseOptions} [options] The formatting options.
  * @return {Promise<ColumnTable>} A Promise to an Arquero table.
  */
-export async function parseCSV(input, options) {
-  const stream = isString(input) ? toTextStream(input)
-    : input instanceof ReadableStream ? input
-    : error('parseCSV input must be a string or ReadableStream');
-  const { columns, names } = await streamCSV(stream, options);
-  return new ColumnTable(columns, names);
+export async function fromCSVStream(stream, options) {
+  return parseTextRowsStream(stream, transforms(options), options);
 }
 
-function streamCSV(input, {
+/**
+ * Load a CSV file from a URL and return a Promise for an Arquero table.
+ * @param {string} path The URL or file path to load.
+ * @param {import('./types.js').LoadOptions & CSVParseOptions} [options]
+ *  CSV parse options.
+ * @return {Promise<ColumnTable>} A Promise to an Arquero table.
+ * @example aq.loadCSV('data/table.csv')
+ * @example aq.loadTSV('data/table.tsv', { delimiter: '\t' })
+ */
+export async function loadCSV(path, options) {
+  return fromCSVStream(await textStream(path, options), options);
+}
+
+function transforms({
   delimiter = ',',
   skip = 0,
-  comment = undefined,
-  ...options
+  comment = undefined
 } = {}) {
-  return parseTextRows(
-    pipeline(input, [
-      new DelimitedTextStream(delimiter),
-      lineFilter(skip, comment, row => row[0])
-    ]),
-    options
-  );
+  return [
+    delimitedTextTransformer(delimiter),
+    lineFilterTransformer(skip, comment, row => row[0])
+  ];
 }
