+++ src/format/stream/fixed-text-stream.js
@@ -1,21 +1,18 @@
 /**
- * @extends {TransformStream<string[],string[][]>}
+ * Returns a fixed text stream transformer.
+ * @param {[number, number][]} positions The fixed positions delimiting records
+ * @returns {Transformer<string[], string[][]>}
  */
-export class FixedTextStream extends TransformStream {
-  /**
-   * @param {[number, number][]} positions
-   */
-  constructor(positions) {
-    super({
-      start() {}, // no-op
-      flush() {}, // no-op
-      transform(chunk, controller) {
-        controller.enqueue(
-          chunk.map(
-            line => positions.map(([i, j]) => line.slice(i, j).trim())
-          )
-        );
-      }
-    });
-  }
+export function fixedTextTransformer(positions) {
+  return {
+    start() {}, // no-op
+    flush() {}, // no-op
+    transform(chunk, controller) {
+      controller.enqueue(
+        chunk.map(
+          line => positions.map(([i, j]) => line.slice(i, j).trim())
+        )
+      );
+    }
+  };
 }
