+++ src/format/stream/line-filter-stream.js
@@ -6,11 +6,11 @@ import { identity } from '../../util/identity.js';
  * @param {number} [skip]
  * @param {string} [comment]
  * @param {(value: T) => string} [accessor]
- * @returns {LineFilterStream<T[]> | null}
+ * @returns {TransformStream<T[]> | null}
  */
-export function lineFilter(skip, comment, accessor = identity) {
-  const drop = shouldDrop(skip, comment, accessor);
-  return drop ? new LineFilterStream(drop) : null;
+export function lineFilter(skip, comment, accessor) {
+  const transform = lineFilterTransformer(skip, comment, accessor);
+  return transform ? new TransformStream(transform) : null;
 }
 
 function shouldDrop(skip, comment, accessor) {
@@ -22,31 +22,45 @@ function shouldDrop(skip, comment, accessor) {
     : null;
 }
 
+// /**
+//  * @template T
+//  * @extends {TransformStream<T[],T[]>}
+//  */
+// export class LineFilterStream extends TransformStream {
+//   /**
+//    * @param {(line: T, index: number) => boolean} drop
+//    */
+//   constructor(drop) {
+//     super(lineFilterTransformer(drop));
+//   }
+// }
+
 /**
+ * Returns a new line filter stream transformer.
  * @template T
- * @extends {TransformStream<T[],T[]>}
+ * @param {number} [skip]
+ * @param {string} [comment]
+ * @param {(value: T) => string} [accessor]
+ * @returns {Transformer<T[], T[]>}
  */
-export class LineFilterStream extends TransformStream {
-  /**
-   * @param {(line: T, index: number) => boolean} drop
-   */
-  constructor(drop) {
-    let i = 0;
-    super({
-      start() {}, // no-op
-      flush() {}, // no-op
-      transform(chunk, controller) {
-        const n = chunk.length;
-        const bits = new BitSet(n);
-        for (let c = 0; c < chunk.length; ++c, ++i) {
-          if (drop(chunk[c], i)) bits.set(c);
-        }
-        controller.enqueue(
-          bits.count()
-            ? chunk.filter((_, c) => !bits.get(c))
-            : chunk
-        );
+export function lineFilterTransformer(skip, comment, accessor = identity) {
+  const drop = shouldDrop(skip, comment, accessor);
+  if (!drop) return null;
+  let i = 0;
+  return {
+    start() {}, // no-op
+    flush() {}, // no-op
+    transform(chunk, controller) {
+      const n = chunk.length;
+      const bits = new BitSet(n);
+      for (let c = 0; c < chunk.length; ++c, ++i) {
+        if (drop(chunk[c], i)) bits.set(c);
       }
-    });
-  }
+      controller.enqueue(
+        bits.count()
+          ? chunk.filter((_, c) => !bits.get(c))
+          : chunk
+      );
+    }
+  };
 }
